# account/docs/swagger.yaml
definitions:
  body.CreateAccount:
    properties:
      email:
        example: test@gmail.com
        type: string
      password:
        example: password
        type: string
    required:
    - email
    type: object
  body.UpdateAccount:
    properties:
      password:
        example: password
        type: string
    type: object
  model.Account:
    properties:
      accessToken:
        example: accesstoken
        type: string
      createdAt:
        example: "2019-12-23 12:27:37"
        type: string
      email:
        example: test@gmail.com
        type: string
      id:
        example: accountId
        type: string
      updatedAt:
        example: "2019-12-23 12:27:37"
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
  /accounts:
    delete:
      description: delete account
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Account'
      security:
      - AccessToken: []
      tags:
      - Account
    get:
      consumes:
      - application/json
      parameters:
      - description: account email
        in: query
        name: email
        type: string
      - description: account password
        in: query
        name: password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Account'
      security:
      - AccessToken: []
      tags:
      - Account
    post:
      consumes:
      - application/json
      description: create account
      parameters:
      - description: Create Account data
        in: body
        name: CreateAccount
        required: true
        schema:
          $ref: '#/definitions/body.CreateAccount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Account'
      tags:
      - Account
    put:
      consumes:
      - application/json
      description: update account
      parameters:
      - description: Update Account data
        in: body
        name: UpdateAccount
        required: true
        schema:
          $ref: '#/definitions/body.UpdateAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Account'
      security:
      - AccessToken: []
      tags:
      - Account
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"